chore: extrac 'isRootCypress'
