+++ lib/rules/unsafe-to-chain-command.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const { basename } = require('path')
+const { isRootCypress } = require('./utils/is-root-cypress')
 
 const NAME = basename(__dirname)
 const DESCRIPTION = 'disallow actions within chains'
@@ -101,28 +102,6 @@ module.exports = {
       },
     }
   },
-}
-
-/**
- * @param {import('estree').Node} node
- * @returns {boolean}
- */
-const isRootCypress = (node) => {
-  if (
-    node.type !== 'CallExpression'
-    || node.callee.type !== 'MemberExpression'
-  ) {
-    return false
-  }
-
-  if (
-    node.callee.object.type === 'Identifier'
-    && node.callee.object.name === 'cy'
-  ) {
-    return true
-  }
-
-  return isRootCypress(node.callee.object)
 }
 
 /**
