+++ lib/rules/utils/is-root-cypress.js
@@ -1,0 +1,23 @@
+/**
+ * @param {import('estree').Node} node
+ * @returns {boolean}
+ */
+const isRootCypress = (node) => {
+  if (
+    node.type !== 'CallExpression'
+    || node.callee.type !== 'MemberExpression'
+  ) {
+    return false
+  }
+
+  if (
+    node.callee.object.type === 'Identifier'
+    && node.callee.object.name === 'cy'
+  ) {
+    return true
+  }
+
+  return isRootCypress(node.callee.object)
+}
+
+module.exports = { isRootCypress }
